move UserInfoProvider to security package
