+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/services/impl/Be5EventDbLogger.java
@@ -2,7 +2,7 @@ package com.developmentontheedge.be5.modules.core.services.impl;
 
 import com.developmentontheedge.be5.base.lifecycle.Start;
 import com.developmentontheedge.be5.base.services.ProjectProvider;
-import com.developmentontheedge.be5.base.services.UserInfoProvider;
+import com.developmentontheedge.be5.base.security.UserInfoProvider;
 import com.developmentontheedge.be5.databasemodel.DatabaseModel;
 import com.developmentontheedge.be5.metadata.Features;
 import com.developmentontheedge.be5.metadata.model.Query;
