+++ query/src/test/java/com/developmentontheedge/be5/query/QueryBe5ProjectDBTest.java
@@ -2,13 +2,11 @@ package com.developmentontheedge.be5.query;
 
 import com.developmentontheedge.be5.base.BaseModule;
 import com.developmentontheedge.be5.base.services.CoreUtils;
-import com.developmentontheedge.be5.base.services.UserInfoProvider;
 import com.developmentontheedge.be5.database.DatabaseModule;
 import com.developmentontheedge.be5.test.BaseTest;
 import com.developmentontheedge.be5.test.mocks.CoreUtilsForTest;
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
-import com.google.inject.Scopes;
 import com.google.inject.util.Modules;
