+++ server/src/main/java/com/developmentontheedge/be5/server/services/DocumentOperationsPlugin.java
@@ -1,7 +1,7 @@
 package com.developmentontheedge.be5.server.services;
 
 import com.developmentontheedge.be5.base.services.UserAwareMeta;
-import com.developmentontheedge.be5.base.services.UserInfoProvider;
+import com.developmentontheedge.be5.base.security.UserInfoProvider;
 import com.developmentontheedge.be5.metadata.model.Operation;
 import com.developmentontheedge.be5.metadata.model.OperationSet;
 import com.developmentontheedge.be5.metadata.model.Query;
