+++ server/src/main/java/com/developmentontheedge/be5/server/services/impl/FormGeneratorImpl.java
@@ -3,7 +3,7 @@ package com.developmentontheedge.be5.server.services.impl;
 import com.developmentontheedge.be5.base.FrontendConstants;
 import com.developmentontheedge.be5.base.exceptions.Be5Exception;
 import com.developmentontheedge.be5.base.services.UserAwareMeta;
-import com.developmentontheedge.be5.base.services.UserInfoProvider;
+import com.developmentontheedge.be5.base.security.UserInfoProvider;
 import com.developmentontheedge.be5.base.util.HashUrl;
 import com.developmentontheedge.be5.base.util.LayoutUtils;
 import com.developmentontheedge.be5.operation.model.Operation;
